finishing splitting out helpers from main link-checker test
